requirements.txt
anyio==3.6.1
appdirs==1.4.3
apturl==0.5.2
asgiref==3.5.2
astroid==2.11.7
attrs==22.1.0
Babel==2.9.1
backports.zoneinfo==0.2.1
bcrypt==3.1.7
beautifulsoup4==4.11.1
blinker==1.4
Brlapi==0.7.0
cached-property==1.5.2
certifi==2019.11.28
cffi==1.15.1
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
dill==0.3.5.1
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
Django==4.0.6
django-allauth==0.51.0
django-cors-headers==3.13.0
django-filter==22.1
django-rest-knox==4.2.0
djangorestframework==3.13.1
docutils==0.16
duplicity==0.8.12.0
ebaysdk==2.1.5
entrypoints==0.3
fail2ban==0.11.1
fastapi==0.79.0
fastapi-pagination==0.9.3
fasteners==0.14.1
feedparser==5.2.1
filelock==3.0.12
freezegun==0.3.15
future==0.18.2
gevent==20.9.0
greenlet==0.4.17
html2text==2016.9.19
httplib2==0.14.0
idna==2.8
importlib-metadata==1.5.0
isodate==0.6.1
isort==5.10.1
Jinja2==2.11.3
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.7.1
libsass==0.18.0
lockfile==0.12.2
louis==3.12.0
lxml==4.6.5
Mako==1.0.4
MarkupSafe==1.1.0
mccabe==0.7.0
mock==2.0.0
monotonic==1.5
more-itertools==4.2.0
netifaces==0.10.4
nftables==0.1
num2words==0.5.6
oauthlib==3.1.0
ofxparse==0.19
paramiko==2.6.0
passlib==1.7.3
pexpect==4.6.0
Pillow==9.0.1
platformdirs==2.5.2
polib==1.1.0
psutil==5.6.7
psycopg2==2.8.6
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycparser==2.21
pycups==1.9.73
pydantic==1.9.1
pydot==1.4.1
PyGObject==3.36.0
pyinotify==0.9.6
PyJWT==1.7.1
pylint==2.14.5
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==3.0.9
PyPDF2==1.26.0
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.8
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-ldap==3.4.0
python-stdnum==1.13
python3-openid==3.2.0
pytz==2019.3
pyusb==1.0.2
pyxdg==0.26
PyYAML==5.3.1
qrcode==6.1
reportlab==3.5.59
requests==2.25.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
requests-unixsocket==0.2.0
SecretStorage==2.3.1
selinux==3.0
simplejson==3.16.0
six==1.14.0
slip==0.6.5
slip.dbus==0.6.5
sniffio==1.2.0
soupsieve==2.3.2.post1
South==1.0.2
SQLAlchemy==1.4.39
SQLAlchemy-ImageAttach==1.1.0
SQLAlchemy-Utils==0.38.3
sqlalchemy2-stubs==0.0.2a24
sqlmodel==0.0.6
sqlparse==0.4.2
ssh-import-id==5.10
starlette==0.19.1
stripe==4.1.0
systemd-python==234
testresources==2.0.1
tomli==2.0.1
tomlkit==0.11.1
typing_extensions==4.3.0
ubuntu-advantage-tools==27.11.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
virtualenv==20.0.17
vobject==0.9.6.1
wadllib==1.3.3
Wand==0.6.8
Werkzeug==0.16.1
wrapt==1.14.1
xkit==0.0.0
xlrd==1.2.0
XlsxWriter==1.1.2
xlwt==1.3.0
zeep==3.4.0
zipp==1.0.0
zope.event==4.5.0
zope.interface==5.4.0
